Organize frontend components and pages
