+++ frontend/src/components/LoginForm.tsx
@@ -2,8 +2,8 @@ import { useForm } from 'react-hook-form'
 import { useState } from 'react'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { login, storeToken } from './api'
-import type { LoginData } from './api'
+import { login, storeToken } from '../api'
+import type { LoginData } from '../api'
 import { useToast } from './ToastProvider'
 import RoleSlider from './RoleSlider'
 import type { Role } from './RoleSlider'
