+++ frontend/src/components/RoleSlider.tsx
