+++ frontend/src/pages/CallsPage.tsx
@@ -1,4 +1,4 @@
-import CallList from './CallList'
+import CallList from '../components/CallList'
 
 export default function CallsPage() {
   return <CallList />
