+++ frontend/src/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from './LoginForm'
+import LoginForm from '../components/LoginForm'
 
 export default function LoginPage() {
   return (
